WT-17972 Rename header to schema_disagg_abort.h
